Remove dataset slicing tests to their own file.
